libctf, include: API review: delete ctf_dict_open_sections et al
This function has no practical use; it lets you open an archive but also
associate it with a string and symbol table different from that in the
ELF file it was linked to.  It might in theory be useful if standalone
archives not in ELF sections but associated with ELF string and symbol
tables existed, but they don't, and why would they?  Standalone archives
are by definition not associated with an ELF file, so are hardly going
to reference some ELF file's strtab and symtab.

Delete the obsolete alias ctf_arc_open_by_name_sections too.

This lets us fold ctf_dict_open_sections into its only remaining caller.