libctf, include: API review: extend ctf_type_set_conflicting
We'd like the libctf API to make common things easy and not to be
cluttered up with rarely-needed things except when those things are
actually required.  The current API's ctf_add_*() calls are all far from
that ideal because every single one of them takes a badly-named "flag"
which indicates whether this type is root-visible or not.

Almost all types are root-visible: C doesn't even have a concept of a
non-root-visible type, and BTF doesn't implement it at all.  CTFv4 uses
the concept solely for conflicting types that have nonetheless ended up
in the same dict (usually because several dicts have been merged into
one via the CU-mapping mechanism).  Normal links cannot do this, so
almost no ctf_add_*() callers are ever going to want a non-root-visible
type.  So we should definitely drop the flag...

... but if we do, how does code that does want a conflicting,
non-root-visible type indicate it? It can't pass CTF_ADD_NONROOT if that
is deleted, but it also can't use ctf_type_set_conflicting because that
can only be called on existing types, but if a type is conflicting its
name is going to clash with an existing type, so we're going to want to
mark it as such from the moment it's added, before its type ID is even
assigned (at the time name-clash detection is carried out).

Fix this by extending ctf_type_set_conflicting to let you say that type
0 is conflicting, which is taken to mean that the next
successfully-added type in this dict is marked conflicting as soon as it
is added, just like passing CTF_ADD_NONROOT currently does.  Type 0 has
no physical existence (BTF uses it for void), so it cannot be
conflicting and so is free in this API for this obscure edge-case use.

(This interface would be horrible if it was meant for frequent use, but
only deduplicators are likely to use it, and not most of those.)